package-private getCache()
