+++ platform-sdk/swirlds-virtualmap/src/test/java/com/swirlds/virtualmap/VirtualMapTests.java
@@ -33,6 +33,7 @@ import com.swirlds.metrics.api.Metrics;
 import com.swirlds.virtualmap.datasource.VirtualDataSourceBuilder;
 import com.swirlds.virtualmap.datasource.VirtualLeafBytes;
 import com.swirlds.virtualmap.internal.VirtualMapStatistics;
+import com.swirlds.virtualmap.internal.cache.VirtualNodeCache;
 import com.swirlds.virtualmap.test.fixtures.TestKey;
 import com.swirlds.virtualmap.test.fixtures.TestValue;
 import com.swirlds.virtualmap.test.fixtures.TestValueCodec;
@@ -46,9 +47,11 @@ import java.time.Duration;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Random;
+import java.util.Set;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
+import java.util.stream.Collectors;
 import org.hiero.base.crypto.Hash;
 import org.hiero.base.exceptions.ReferenceCountException;
 import org.hiero.base.utility.test.fixtures.file.TestFileSystemManager;
@@ -1250,5 +1253,139 @@ class VirtualMapTests extends VirtualTestBase {
     @Test
     void getVersion() {
         assertEquals(4, createMap().getVersion());
+    }
+
+    @Test
+    @Tags({@Tag("VirtualMerkle"), @Tag("VirtualNodeCache"), @Tag("Leaf")})
+    @DisplayName("deletedLeaves()")
+    void deletedLeaves() {
+        // CREATED followed by UPDATED, UPDATED+DELETED, DELETED
+        // CREATED+UPDATED followed by UPDATED, UPDATED+DELETED, DELETED
+        // UPDATED followed by UPDATED, UPDATED+DELETED, DELETED
+        // DELETED followed by CREATED, CREATED+UPDATED, CREATED+DELETED, CREATED+UPDATED+DELETED, DELETED (nop)
+
+        // Create the following chain of mutations:
+        // A: [D, v2] -> [U+D (AARDVARK), v1] -> [C (APPLE), v0]
+        // B: [D, v3] -> [C+U (BEAR, BLASTOFF), v2] -> [D, v1] -> [C (BANANA), v0]
+        // C: [C+U+D (CHEMISTRY, CHAD), v3] -> [D, v2] -> [U (COMET), v1] -> [C+U (CHERRY, CUTTLEFISH), v0]
+        // D: [C+U (DISCIPLINE, DENMARK), v2] -> [U+D (DRACO), v1] -> [C+U (DATE, DOG), v0]
+        // E: [C+U (EXOPLANET, ECOLOGY), v3] -> [D, v2] -> [C+U (EGGPLANT, EMU), v0]
+        // F: [C (FORCE), v3] -> [D, v2] -> [U (FOX), v1] -> [C (FIG), v0]
+        // G: [U (GRAVITY), v3] -> [U (GOOSE), v2] -> [C (GRAPE), v1]
+
+        final VirtualMap map0 = createMap();
+        final VirtualNodeCache cache0 = map0.getCache();
+        // A: [C (APPLE), v0]
+        // B: [C (BANANA), v0]
+        // C: [C+U (CHERRY, CUTTLEFISH), v0]
+        // D: [C+U (DATE, DOG), v0]
+        // E: [C+U (EGGPLANT, EMU), v0]
+        // F: [C (FIG), v0]
+        map0.put(A_KEY, APPLE, TestValueCodec.INSTANCE);
+        map0.put(B_KEY, BANANA, TestValueCodec.INSTANCE);
+        map0.put(C_KEY, CHERRY, TestValueCodec.INSTANCE);
+        map0.put(C_KEY, CUTTLEFISH, TestValueCodec.INSTANCE);
+        map0.put(D_KEY, DATE, TestValueCodec.INSTANCE);
+        map0.put(D_KEY, DOG, TestValueCodec.INSTANCE);
+        map0.put(E_KEY, EGGPLANT, TestValueCodec.INSTANCE);
+        map0.put(E_KEY, EMU, TestValueCodec.INSTANCE);
+        map0.put(F_KEY, FIG, TestValueCodec.INSTANCE);
+
+        final VirtualMap map1 = map0.copy();
+        final VirtualNodeCache cache1 = map1.getCache();
+
+        // A: [U+D (AARDVARK), v1]
+        // B: [D, v1]
+        // C: [U (COMET), v1]
+        // D: [U+D (DRACO), v1]
+        // F: [U (FOX), v1]
+        // G: [C (GRAPE), v1]
+        map1.put(A_KEY, AARDVARK, TestValueCodec.INSTANCE);
+        map1.remove(A_KEY);
+        map1.remove(B_KEY);
+        map1.put(C_KEY, COMET, TestValueCodec.INSTANCE);
+        map1.put(D_KEY, DRACO, TestValueCodec.INSTANCE);
+        map1.remove(D_KEY);
+        map1.put(F_KEY, FOX, TestValueCodec.INSTANCE);
+        map1.put(G_KEY, GRAPE, TestValueCodec.INSTANCE);
+
+        final VirtualMap map2 = map1.copy();
+        final VirtualNodeCache cache2 = map2.getCache();
+
+        // A: [D, v2]
+        // B: [C+U (BEAR, BLASTOFF), v2]
+        // C: [D, v2]
+        // D: [C+U (DISCIPLINE, DENMARK), v2]
+        // E: [D, v2]
+        // F: [D, v2]
+        // G: [U (GOOSE), v2]
+        map2.remove(A_KEY, TestValueCodec.INSTANCE);
+        map2.put(B_KEY, BEAR, TestValueCodec.INSTANCE);
+        map2.put(B_KEY, BLASTOFF, TestValueCodec.INSTANCE);
+        map2.remove(C_KEY);
+        map2.put(D_KEY, DISCIPLINE, TestValueCodec.INSTANCE);
+        map2.put(D_KEY, DENMARK, TestValueCodec.INSTANCE);
+        map2.remove(E_KEY);
+        map2.remove(F_KEY);
+        map2.put(G_KEY, GOOSE, TestValueCodec.INSTANCE);
+
+        final VirtualMap map3 = map2.copy();
+        final VirtualNodeCache cache3 = map3.getCache();
+
+        // B: [D, v3]
+        // C: [C+U+D (CHEMISTRY, CHAD), v3]
+        // E: [C+U (EXOPLANET, ECOLOGY), v3]
+        // F: [C (FORCE), v3]
+        // G: [U (GRAVITY), v3]
+        map3.remove(B_KEY);
+        map3.put(C_KEY, CHEMISTRY, TestValueCodec.INSTANCE);
+        map3.put(C_KEY, CHAD, TestValueCodec.INSTANCE);
+        map3.remove(C_KEY);
+        map3.put(E_KEY, EXOPLANET, TestValueCodec.INSTANCE);
+        map3.put(E_KEY, ECOLOGY, TestValueCodec.INSTANCE);
+        map3.put(F_KEY, FORCE, TestValueCodec.INSTANCE);
+        map3.put(G_KEY, GRAVITY, TestValueCodec.INSTANCE);
+
+        // One last copy, so we can get the dirty leaves without an exception
+        final VirtualMap map4 = map3.copy();
+
+        final List<VirtualLeafBytes> deletedLeaves0 = cache0.deletedLeaves().toList();
+        assertEquals(0, deletedLeaves0.size(), "No deleted leaves in cache0");
+
+        cache0.seal();
+        cache1.seal();
+        cache0.merge();
+        validateDeletedLeaves(
+                cache1.deletedLeaves().collect(Collectors.toList()), Set.of(A_KEY, B_KEY, D_KEY), "cache1");
+
+        cache2.seal();
+        cache1.merge();
+        validateDeletedLeaves(
+                cache2.deletedLeaves().collect(Collectors.toList()), Set.of(A_KEY, C_KEY, E_KEY, F_KEY), "cache2");
+
+        cache3.seal();
+        cache2.merge();
+        validateDeletedLeaves(
+                cache3.deletedLeaves().collect(Collectors.toList()), Set.of(A_KEY, B_KEY, C_KEY), "cache3");
+
+        map0.release();
+        map1.release();
+        map2.release();
+        map3.release();
+        map4.release();
+    }
+
+    private void validateDeletedLeaves(
+            final List<VirtualLeafBytes> deletedLeaves, final Set<Bytes> expectedKeys, final String name) {
+
+        assertEquals(expectedKeys.size(), deletedLeaves.size(), "Not enough deleted leaves in " + name);
+
+        final Set<Bytes> keys =
+                deletedLeaves.stream().map(VirtualLeafBytes::keyBytes).collect(Collectors.toSet());
+        assertEquals(deletedLeaves.size(), keys.size(), "Two records with the same key exist in " + name);
+
+        for (final var rec : deletedLeaves) {
+            assertTrue(keys.remove(rec.keyBytes()), "A record does not have the expected key in " + name);
+        }
     }
 }
